// packages/@ros-cdk/services/slb/lib/slb.generated.ts
    export interface BackendServersProperty {
        /**
         * @Property type: The backend server type. Valid values:ecs: ECS instance (default)eni: Elastic Network Interface (ENI)
         */
        readonly type?: string;
        /**
         * @Property serverId: Need one valid instance id. The instance status should running.
         */
        readonly serverId: string;
        /**
         * @Property description: A description of the backend server.
         */
        readonly description?: string;
        /**
         * @Property serverIp: The IP of the instance.
         */
        readonly serverIp?: string;
        /**
         * @Property weight: The weight of backend server of load balancer. From 0 to 100, 0 means offline. Default is 100.
         */
        readonly weight: number;
    }
// ...
export interface RosVServerGroupProps {

    /**
     * @Property loadBalancerId: The id of load balancer.
     */
    readonly loadBalancerId: string;

    /**
     * @Property vServerGroupName: Display name of the VServerGroup.
     */
    readonly vServerGroupName: string;

    /**
     * @Property backendServers: The list of a combination of ECS Instance-Port-Weight.Same ecs instance with different port is allowed, but same ecs instance with same port isn't.
     */
    readonly backendServers?: Array<RosVServerGroup.BackendServersProperty | ros.IResolvable> | ros.IResolvable;
}
// ...
export class RosVServerGroup extends ros.RosResource {
    /**
     * The resource type name for this resource class.
     */
    public static readonly ROS_RESOURCE_TYPE_NAME = "ALIYUN::SLB::VServerGroup";

    /**
     * A factory method that creates a new instance of this class from an object
     * containing the properties of this ROS resource.
     */

    /**
     * @Attribute BackendServers: Backend server list in this VServerGroup.
     */
    public readonly attrBackendServers: any;

    /**
     * @Attribute VServerGroupId: The id of VServerGroup created.
     */
    public readonly attrVServerGroupId: any;

    public enableResourcePropertyConstraint: boolean;


    /**
     * @Property loadBalancerId: The id of load balancer.
     */
    public loadBalancerId: string;

    /**
     * @Property vServerGroupName: Display name of the VServerGroup.
     */
    public vServerGroupName: string;

    /**
     * @Property backendServers: The list of a combination of ECS Instance-Port-Weight.Same ecs instance with different port is allowed, but same ecs instance with same port isn't.
     */
    public backendServers: Array<RosVServerGroup.BackendServersProperty | ros.IResolvable> | ros.IResolvable | undefined;

    /**
     * Create a new `ALIYUN::SLB::VServerGroup`.
     *
     * @param scope - scope in which this resource is defined
     * @param id    - scoped id of the resource
     * @param props - resource properties
     */
    constructor(scope: ros.Construct, id: string, props: RosVServerGroupProps, enableResourcePropertyConstraint: boolean) {
        super(scope, id, { type: RosVServerGroup.ROS_RESOURCE_TYPE_NAME, properties: props });
        this.attrBackendServers = ros.Token.asString(this.getAtt('BackendServers'));
        this.attrVServerGroupId = ros.Token.asString(this.getAtt('VServerGroupId'));

        this.enableResourcePropertyConstraint = enableResourcePropertyConstraint;
        this.loadBalancerId = props.loadBalancerId;
        this.vServerGroupName = props.vServerGroupName;
        this.backendServers = props.backendServers;
    }


    protected get rosProperties(): { [key: string]: any }  {
        return {
            loadBalancerId: this.loadBalancerId,
            vServerGroupName: this.vServerGroupName,
            backendServers: this.backendServers,
        };
    }
    protected renderProperties(props: {[key: string]: any}): { [key: string]: any }  {
        return rosVServerGroupPropsToRosTemplate(props, this.enableResourcePropertyConstraint);
    }
}

export namespace RosVServerGroup {
    /**
     * @stability external
     */
    export interface BackendServersProperty {
        /**
         * @Property serverId: Need one valid ECS instance id.
         */
        readonly serverId: string;
        /**
         * @Property port: The port of backend server. From 1 to 65535.
         */
        readonly port: number;
        /**
         * @Property weight: The weight of backend server of load balancer. From 0 to 100, 0 means offline. Default is 100.
         */
        readonly weight?: number;
    }
}